Move C++ coverage utility from ct/common to lib/cpp
