+++ go/lib/cpp/coverage_test.go
@@ -8,7 +8,7 @@
 // On the date above, in accordance with the Business Source License, use of
 // this software will be governed by the GNU Lesser General Public License v3.
 
-package common
+package cpp
 
 import (
 	"flag"
